requirements.txt
pygame
mutagen
pillow
